refactor: удалены неиспользуемые модули
