+++ readme.md
@@ -97,7 +97,7 @@
 {
   "result": true
 }
-####  READ /resources/id  
+####  GET /resources/id  
  
 Возвращает список всех ресурсов или список ресурсов, соответствующих указанному фильтру.
